pyproject.toml
[project]
name = "deep-search-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai==1.98.0",
    "openai-agents>=0.2.5",
    "tavily-python>=0.7.10",
]
